requirements.txt
librosa
evaluate>=0.30
git+https://github.com/huggingface/transformers
gradio
jiwer
pandas
dataset
unicode_tr
setuptools==59.5.0
more-itertools